// ui/src/app/api/models/budget-report-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BudgetReportAccountRow } from '../models/budget-report-account-row';
export interface BudgetReportResponse {
  accounts?: Array<BudgetReportAccountRow>;
  expenses?: number;
  income?: number;
  lastPayrollAmount?: number | null;
  lastPayrollDate?: string | null;
  netProfit?: number;
  period?: string;
}